requirements.txt
pandas
scikit-learn
joblib
pip
python-dateutil
pytz
scipy
setuptools
six
threadpoolctl
tzdata
wheel
rembg
Pillow
matplotlib
